+++ lib/src/cell_builder.dart
@@ -4,5 +4,5 @@ import 'package:flutter/widgets.dart';
 /// Signature for a function that builds a widget for a given row.
 ///
 /// Used by [EasyTableColumn].
-typedef EasyTableCellBuilder<ROW> = Widget Function(
+typedef DaviCellBuilder<ROW> = Widget Function(
     BuildContext context, RowData<ROW> data);