+++ test/column_grow.dart
@@ -4,15 +4,15 @@ import 'package:flutter_test/flutter_test.dart';
 void main() {
   group('Column', () {
     test('grow', () {
-      EasyTableColumn column = EasyTableColumn();
+      DaviColumn column = DaviColumn();
       expect(column.grow, null);
-      column = EasyTableColumn(grow: null);
+      column = DaviColumn(grow: null);
       expect(column.grow, null);
-      column = EasyTableColumn(grow: -1);
+      column = DaviColumn(grow: -1);
       expect(column.grow, 1);
-      column = EasyTableColumn(grow: 0);
+      column = DaviColumn(grow: 0);
       expect(column.grow, 1);
-      column = EasyTableColumn(grow: 2);
+      column = DaviColumn(grow: 2);
       expect(column.grow, 2);
 
       column.grow = null;